[feature/restx-resources] Refactoring rest_apis
