--- starwars_api/rest_apis/starwars/schemas/__init__.py
@@ -1,5 +1,0 @@
-from flask import Flask
-
-
-def init_app(app: Flask) -> None:  # NOQA
-    ...
