+++ starwars_api/rest_apis/starwars/serializers/movie.py
@@ -1,0 +1,8 @@
+from marshmallow_mongoengine import ModelSchema
+
+from starwars_api.models.starwars.movie import Movie
+
+
+class MovieSchema(ModelSchema):
+    class Meta:
+        model = Movie
