+++ starwars_api/rest_apis/starwars/serializers/planet.py
@@ -1,0 +1,8 @@
+from marshmallow_mongoengine import ModelSchema
+
+from starwars_api.models.starwars.planet import Planet
+
+
+class PlanetSchema(ModelSchema):
+    class Meta:
+        model = Planet
